🚚 move: CarouselContainer Fixes #4
