+++ src/components/layouts/HomeLayout/index.tsx
@@ -6,41 +6,9 @@ import { MainContainer } from "@/components/atoms/containers/MainContainer";
 import { MiddleAreaContainer } from "@/components/atoms/containers/MiddleAreaContainer";
 import { RootLayout } from "@/components/layouts/RootLayout";
 import { CategoryListSide } from "@/components/molecules/CategoryListSide";
+import { CarouselContainer } from "@/components/organisms/CarouselContainer";
 import { useMedia } from "@/hooks/useMedia";
 import type { TCategory, TConfig, TPickup, TRankedArticle } from "@/types";
-
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-expect-error
-const PickupArticleCarousel = dynamic(() =>
-  import("@/components/organisms/PickupArticleCarousel").then((mod) => mod.PickupArticleCarousel)
-);
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-expect-error
-const PopularArticleCarousel = dynamic(() =>
-  import("@/components/organisms/PopularArticleCarousel").then((mod) => mod.PopularArticleCarousel)
-);
-
-type CarouselContainerProps = { pickup: TPickup; popularArticles: TRankedArticle[] };
-
-const CarouselContainer: React.FC<CarouselContainerProps> = ({ pickup, popularArticles }) => {
-  if (pickup.articles.length > 0) {
-    return (
-      <div className="px-0 mt-8 md:px-16 md:mt-16">
-        <PickupArticleCarousel pickup={pickup} />
-      </div>
-    );
-  }
-
-  if (popularArticles.length > 0) {
-    return (
-      <div className="px-0 mt-8 md:px-16 md:mt-16">
-        <PopularArticleCarousel articles={popularArticles} />
-      </div>
-    );
-  }
-
-  return null;
-};
 
 type HomeLayoutProps = {
   children: React.ReactNode;
